Added Bundles Removed ROMS
Why we removed roms?
roms are basically zip files but with an other name this is redundant

Why we then Added Bundles?
Because of the latest update "bd981ef55b2e023af1fc2d3e91fe8ba0aaee3c78" in which we finally added externals to make transportion and sharing of XBins with external dependencies easier using only "xbd bundle some_file.xbin external.so some_file.xbin" you will get a package with the dependencie and the xbin then to run you only need to call "xbd run xbd.bundle" and it automatically runs but warning never use run in same directory because it overides existing files and in result deletes them